Move cpu struct and functions to cpu.h, move proc.c declarations from defs.h to proc.h
