+++ kernel/cpu.c
@@ -1,0 +1,25 @@
+#include "cpu.h"
+
+#include "riscv.h"
+
+struct cpu cpus[NCPU];
+
+// Must be called with interrupts disabled,
+// to prevent race with process being moved
+// to a different CPU.
+int
+cpuid()
+{
+  int id = r_tp();
+  return id;
+}
+
+// Return this CPU's cpu struct.
+// Interrupts must be disabled.
+struct cpu*
+mycpu(void) {
+  int id = cpuid();
+  struct cpu *c = &cpus[id];
+  return c;
+}
+
