+++ kernel/proc.c
@@ -1,12 +1,12 @@
+#include "proc.h"
+
+#include "cpu.h"
 #include "types.h"
 #include "param.h"
 #include "memlayout.h"
 #include "riscv.h"
 #include "spinlock.h"
-#include "proc.h"
 #include "defs.h"
-
-struct cpu cpus[NCPU];
 
 struct proc proc[NPROC];
 
@@ -40,25 +40,6 @@ procinit(void)
       p->kstack = va;
   }
   kvminithart();
-}
-
-// Must be called with interrupts disabled,
-// to prevent race with process being moved
-// to a different CPU.
-int
-cpuid()
-{
-  int id = r_tp();
-  return id;
-}
-
-// Return this CPU's cpu struct.
-// Interrupts must be disabled.
-struct cpu*
-mycpu(void) {
-  int id = cpuid();
-  struct cpu *c = &cpus[id];
-  return c;
 }
 
 // Return the current struct proc *, or zero if none.
